feat(tests): move unit tests for deprecated endpoints to own file
